// web/src/lib/api-types.ts
export interface components {
    schemas: {
        ErrorDetail: {
            /** @description Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id' */
            location?: string;
            /** @description Error message text */
            message?: string;
            /** @description The value at the given location */
            value?: unknown;
        };
        ErrorModel: {
            /**
             * Format: uri
             * @description A URL to the JSON Schema for this object.
             * @example https://example.com/schemas/ErrorModel.json
             */
            readonly $schema?: string;
            /**
             * @description A human-readable explanation specific to this occurrence of the problem.
             * @example Property foo is required but is missing.
             */
            detail?: string;
            /** @description Optional list of individual error details */
            errors?: components["schemas"]["ErrorDetail"][] | null;
            /**
             * Format: uri
             * @description A URI reference that identifies the specific occurrence of the problem.
             * @example https://example.com/error-log/abc123
             */
            instance?: string;
            /**
             * Format: int64
             * @description HTTP status code
             * @example 400
             */
            status?: number;
            /**
             * @description A short, human-readable summary of the problem type. This value should not change between occurrences of the error.
             * @example Bad Request
             */
            title?: string;
            /**
             * Format: uri
             * @description A URI reference to human-readable documentation for the error.
             * @default about:blank
             * @example https://example.com/errors/example
             */
            type: string;
        };
        "Get-healthResponse": {
            /**
             * Format: uri
             * @description A URL to the JSON Schema for this object.
             * @example https://example.com/schemas/Get-healthResponse.json
             */
            readonly $schema?: string;
            message: string;
        };
        GetPresignedURLOutputBody: {
            /**
             * Format: uri
             * @description A URL to the JSON Schema for this object.
             * @example https://example.com/schemas/GetPresignedURLOutputBody.json
             */
            readonly $schema?: string;
            url: string;
        };
        Image: {
            /** Format: date-time */
            created_at: string;
            id: string;
            image_key: string;
            task_id: string;
        };
        Job: {
            /**
             * Format: uri
             * @description A URL to the JSON Schema for this object.
             * @example https://example.com/schemas/Job.json
             */
            readonly $schema?: string;
            /** Format: date-time */
            created_at: string;
            id: string;
            tasks: components["schemas"]["Task"][] | null;
        };
        Task: {
            /** Format: date-time */
            created_at: string;
            id: string;
            input_image: components["schemas"]["Image"];
            job_id: string;
            options: string;
            output_image: components["schemas"]["Image"];
            status: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}